workspace/pyproject.toml
[project]
name = "infinite-stock-trader"
version = "0.1.0"
description = "Automatic Infinite Stock Trader using Dollar-Cost Averaging algorithm"
authors = [
    {name = "James Yoo"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0.0"
dependencies = [
    "alpaca-py (>=0.43.2,<0.44.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "toml (>=0.10.2,<0.11.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
